version.intValue を見るべき

+++ core/ui/src/main/java/blue/starry/mitsubachi/ui/locale/LocaleAware.kt
@@ -21,7 +21,7 @@ fun <T> LocaleAware(
     }
   }
 
-  return key(version) {
+  return key(version.intValue) {
     block()
   }
 }
